Basic Functions

OptControl.generateJuMPcodes — Function

generateJuMPcodes(L, F, state, u, tspan, t0, tf)
generateJuMPcodes(L, F, state, u, tspan, t0, tf, Φ)
generateJuMPcodes(L, F, state, u, tspan, t0, tf, Φ, tf_constraint; state_ub, state_lb, u_ub, u_lb, N, discretization, model_name, writeFilePath, tolerance)

generateJuMPcodes will generate the solution code that using JuMP.jl to solve problem

A general form of optimal control problem:

\[min (\Phi(\boldsymbol{x}(t_f),t_f)+\int_{t_0}^{t_f} L[\boldsymbol{x}(t),\boldsymbol{u}(t),t]dt) \\ s.t. \hspace{0.4cm} \dot{\boldsymbol{x}} =f[\boldsymbol{x}(t),\boldsymbol{u}(t),t]\]

args:

L: above L

\[L[\boldsymbol{x}(t),\boldsymbol{u}(t),t]\]

F: above f

\[f[\boldsymbol{x}(t),\boldsymbol{u}(t),t]\]

state: states(variable) in F

u: control variable u in F

tspan: time field

t0: initial value at tspan[1], length of t0 must be equal to length of state

tf: final(end) value at tspan[2], length of tf must be equal to length of state

Φ: above Φ, default: nothing

\[\Phi(\boldsymbol{x}(t_f),t_f)\]

tf_constraint: special requirements for end time, default: nothing. Example:

\[x_{1}+x_{2}=0\]

state_ub: state's upper limit, length of state_ub must be equal to length of state, default: nothing

state_lb: state's lower limit, length of state_lb must be equal to length of state, default: nothing

u_ub: u's upper limit, length of state_ub must be equal to length of u, default: nothing

u_lb: u's lower limit, length of state_lb must be equal to length of u, default: nothing

N: Number of discrete, default: 1000

\[dt = (end_time - start_time) / N\]

discretization: discretization methods, default: "trapezoidal"

model_name: name of JuMP model, default: "model"

writeFilePath: path of generated code, default: nothing

tolerance: acceptable_tol of Ipopt, default: 1.0e-6

Example:

using Symbolics,OptControl
@variables t u x[1:2]
f = [0 1; 0 0] * x + [0, 1] * u
L = 0.5 * u^2
t0 = [1.0, 1.0]
tf = [0.0, 0.0]
tspan = (0.0, 2.0)
N = 100
sol = generateJuMPcodes(L, f, x, u, tspan, t0, tf; N=N)

OptControl.generateDEcodes — Function

generateDEcodes(L, F, state, u, tspan, t0, tf) -> Any
generateDEcodes(L, F, state, u, tspan, t0, tf, Φ) -> Any
generateDEcodes(L, F, state, u, tspan, t0, tf, Φ, dt) -> Any
generateDEcodes(L, F, state, u, tspan, t0, tf, Φ, dt, guess) -> Any
generateDEcodes(L, F, state, u, tspan, t0, tf, Φ, dt, guess, writeFilePath) -> Any

generateDEcodes will generate the solution code that using DifferentialEquations.jl to solve problem

A general form of optimal control problem:

\[min (\Phi(\boldsymbol{x}(t_f),t_f)+\int_{t_0}^{t_f} L[\boldsymbol{x}(t),\boldsymbol{u}(t),t]dt) \\ s.t. \hspace{0.4cm} \dot{\boldsymbol{x}} =f[\boldsymbol{x}(t),\boldsymbol{u}(t),t]\]

args:

L: above L

\[L[\boldsymbol{x}(t),\boldsymbol{u}(t),t]\]

F: above f

\[f[\boldsymbol{x}(t),\boldsymbol{u}(t),t]\]

state: states(variable) in F

u: control variable u in F

tspan: time field

t0: initial value at tspan[1], length of t0 must be equal to length of state

tf: final(end) value at tspan[2], length of tf must be equal to length of state

Φ: above Φ, default: nothing

\[\Phi(\boldsymbol{x}(t_f),t_f)\]

dt: dt of BVProblem, default: 0.1

guess: initial value of BVProblem, default: 1.0. length of guess must be equal to (2*length(state)+length(u))

writeFilePath: path of generated code, default: nothing

Example:

using Symbolics,OptControl
@variables t u x[1:2]
f = [0 1; 0 0] * x + [0, 1] * u
L = 0.5 * u^2
t0 = [1.0, 1.0]
tf = [0.0, 0.0]
tspan = (0.0, 2.0)
sol = generateDEcodes(L, f, x, u, tspan, t0, tf)